+++ src/lib.rs
@@ -63,39 +63,39 @@ use std::path::PathBuf;
 /// Runs the Yul mode for the EraVM target.
 ///
 pub fn yul_to_eravm(
-    input_files: &[PathBuf],
-    solc: Option<String>,
+    paths: &[PathBuf],
+    libraries: Vec<String>,
+    solc_path: Option<String>,
     optimizer_settings: era_compiler_llvm_context::OptimizerSettings,
     is_system_mode: bool,
     include_metadata_hash: bool,
     debug_config: Option<era_compiler_llvm_context::DebugConfig>,
 ) -> anyhow::Result<EraVMBuild> {
-    let path = match input_files.len() {
-        1 => input_files.first().expect("Always exists"),
-        0 => anyhow::bail!("The input file is missing"),
-        length => anyhow::bail!(
-            "Only one input file is allowed in the Yul mode, but found {}",
-            length,
-        ),
-    };
-
-    let solc_validator = if is_system_mode {
-        None
-    } else {
-        let mut solc = SolcCompiler::new(
-            solc.unwrap_or_else(|| SolcCompiler::DEFAULT_EXECUTABLE_NAME.to_owned()),
-        )?;
-        if solc.version()?.default != SolcCompiler::LAST_SUPPORTED_VERSION {
-            anyhow::bail!(
-                "The Yul mode is only supported with the latest supported version of the Solidity compiler: {}",
-                SolcCompiler::LAST_SUPPORTED_VERSION,
-            );
+    let mut solc_compiler = match solc_path {
+        Some(_solc_path) if is_system_mode => {
+            anyhow::bail!("The `solc` validator cannot be used when EraVM extensions are enabled");
         }
-
-        Some(solc)
+        Some(solc_path) => {
+            let mut solc = SolcCompiler::new(solc_path)?;
+            if solc.version()?.default != SolcCompiler::LAST_SUPPORTED_VERSION {
+                anyhow::bail!(
+                    "Yul can only be validated with the latest supported version of the Solidity compiler: {}",
+                    SolcCompiler::LAST_SUPPORTED_VERSION,
+                );
+            }
+            Some(solc)
+        }
+        None => None,
     };
 
-    let project = Project::try_from_yul_path(path, solc_validator.as_ref())?;
+    let libraries = SolcStandardJsonInputSettings::parse_libraries(libraries)?;
+
+    let project = Project::try_from_yul_paths(
+        paths,
+        libraries,
+        solc_compiler.as_mut(),
+        debug_config.as_ref(),
+    )?;
 
     let build = project.compile_to_eravm(
         optimizer_settings,
@@ -112,32 +112,35 @@ pub fn yul_to_eravm(
 /// Runs the Yul mode for the EVM target.
 ///
 pub fn yul_to_evm(
-    input_files: &[PathBuf],
-    solc: Option<String>,
+    paths: &[PathBuf],
+    libraries: Vec<String>,
+    solc_path: Option<String>,
     optimizer_settings: era_compiler_llvm_context::OptimizerSettings,
     include_metadata_hash: bool,
     debug_config: Option<era_compiler_llvm_context::DebugConfig>,
 ) -> anyhow::Result<EVMBuild> {
-    let path = match input_files.len() {
-        1 => input_files.first().expect("Always exists"),
-        0 => anyhow::bail!("The input file is missing"),
-        length => anyhow::bail!(
-            "Only one input file is allowed in the Yul mode, but found {}",
-            length,
-        ),
+    let mut solc_compiler = match solc_path {
+        Some(solc_path) => {
+            let mut solc = SolcCompiler::new(solc_path)?;
+            if solc.version()?.default != SolcCompiler::LAST_SUPPORTED_VERSION {
+                anyhow::bail!(
+                    "Yul can only be validated with the latest supported version of the Solidity compiler: {}",
+                    SolcCompiler::LAST_SUPPORTED_VERSION,
+                );
+            }
+            Some(solc)
+        }
+        None => None,
     };
 
-    let mut solc = SolcCompiler::new(
-        solc.unwrap_or_else(|| SolcCompiler::DEFAULT_EXECUTABLE_NAME.to_owned()),
-    )?;
-    if solc.version()?.default != SolcCompiler::LAST_SUPPORTED_VERSION {
-        anyhow::bail!(
-            "The Yul mode is only supported with the latest supported version of the Solidity compiler: {}",
-            SolcCompiler::LAST_SUPPORTED_VERSION,
-        );
-    }
+    let libraries = SolcStandardJsonInputSettings::parse_libraries(libraries)?;
 
-    let project = Project::try_from_yul_path(path, Some(&solc))?;
+    let project = Project::try_from_yul_paths(
+        paths,
+        libraries,
+        solc_compiler.as_mut(),
+        debug_config.as_ref(),
+    )?;
 
     let build = project.compile_to_evm(optimizer_settings, include_metadata_hash, debug_config)?;
 
@@ -148,22 +151,13 @@ pub fn yul_to_evm(
 /// Runs the LLVM IR mode for the EraVM target.
 ///
 pub fn llvm_ir_to_eravm(
-    input_files: &[PathBuf],
+    paths: &[PathBuf],
     optimizer_settings: era_compiler_llvm_context::OptimizerSettings,
     is_system_mode: bool,
     include_metadata_hash: bool,
     debug_config: Option<era_compiler_llvm_context::DebugConfig>,
 ) -> anyhow::Result<EraVMBuild> {
-    let path = match input_files.len() {
-        1 => input_files.first().expect("Always exists"),
-        0 => anyhow::bail!("The input file is missing"),
-        length => anyhow::bail!(
-            "Only one input file is allowed in the LLVM IR mode, but found {}",
-            length,
-        ),
-    };
-
-    let project = Project::try_from_llvm_ir_path(path)?;
+    let project = Project::try_from_llvm_ir_paths(paths)?;
 
     let build = project.compile_to_eravm(
         optimizer_settings,
@@ -180,21 +174,12 @@ pub fn llvm_ir_to_eravm(
 /// Runs the LLVM IR mode for the EVM target.
 ///
 pub fn llvm_ir_to_evm(
-    input_files: &[PathBuf],
+    paths: &[PathBuf],
     optimizer_settings: era_compiler_llvm_context::OptimizerSettings,
     include_metadata_hash: bool,
     debug_config: Option<era_compiler_llvm_context::DebugConfig>,
 ) -> anyhow::Result<EVMBuild> {
-    let path = match input_files.len() {
-        1 => input_files.first().expect("Always exists"),
-        0 => anyhow::bail!("The input file is missing"),
-        length => anyhow::bail!(
-            "Only one input file is allowed in the LLVM IR mode, but found {}",
-            length,
-        ),
-    };
-
-    let project = Project::try_from_llvm_ir_path(path)?;
+    let project = Project::try_from_llvm_ir_paths(paths)?;
 
     let build = project.compile_to_evm(optimizer_settings, include_metadata_hash, debug_config)?;
 
@@ -205,20 +190,11 @@ pub fn llvm_ir_to_evm(
 /// Runs the EraVM assembly mode.
 ///
 pub fn eravm_assembly(
-    input_files: &[PathBuf],
+    paths: &[PathBuf],
     include_metadata_hash: bool,
     debug_config: Option<era_compiler_llvm_context::DebugConfig>,
 ) -> anyhow::Result<EraVMBuild> {
-    let path = match input_files.len() {
-        1 => input_files.first().expect("Always exists"),
-        0 => anyhow::bail!("The input file is missing"),
-        length => anyhow::bail!(
-            "Only one input file is allowed in the EraVM assembly mode, but found {}",
-            length,
-        ),
-    };
-
-    let project = Project::try_from_eravm_assembly_path(path)?;
+    let project = Project::try_from_eravm_assembly_paths(paths)?;
 
     let optimizer_settings = era_compiler_llvm_context::OptimizerSettings::none();
     let build = project.compile_to_eravm(
@@ -236,9 +212,9 @@ pub fn eravm_assembly(
 /// Runs the standard output mode for EraVM.
 ///
 pub fn standard_output_eravm(
-    input_files: &[PathBuf],
+    paths: &[PathBuf],
     libraries: Vec<String>,
-    solc: &mut SolcCompiler,
+    solc_compiler: &mut SolcCompiler,
     evm_version: Option<era_compiler_common::EVMVersion>,
     solc_optimizer_enabled: bool,
     optimizer_settings: era_compiler_llvm_context::OptimizerSettings,
@@ -253,13 +229,13 @@ pub fn standard_output_eravm(
     suppressed_warnings: Option<Vec<Warning>>,
     debug_config: Option<era_compiler_llvm_context::DebugConfig>,
 ) -> anyhow::Result<EraVMBuild> {
-    let solc_version = solc.version()?;
+    let solc_version = solc_compiler.version()?;
     let solc_pipeline = SolcPipeline::new(&solc_version, force_evmla);
 
     let solc_input = SolcStandardJsonInput::try_from_paths(
         SolcStandardJsonInputLanguage::Solidity,
         evm_version,
-        input_files,
+        paths,
         libraries,
         remappings,
         SolcStandardJsonInputSettingsSelection::new_required(Some(solc_pipeline)),
@@ -288,7 +264,7 @@ pub fn standard_output_eravm(
     }
 
     let libraries = solc_input.settings.libraries.clone().unwrap_or_default();
-    let mut solc_output = solc.standard_json(
+    let mut solc_output = solc_compiler.standard_json(
         solc_input,
         Some(solc_pipeline),
         base_path,
@@ -312,11 +288,12 @@ pub fn standard_output_eravm(
         }
     }
 
-    let project = solc_output.try_to_project_solidity(
+    let project = Project::try_from_solidity_sources(
+        &mut solc_output,
         source_code_files,
         libraries,
         solc_pipeline,
-        &solc_version,
+        solc_compiler,
         debug_config.as_ref(),
     )?;
 
@@ -335,9 +312,9 @@ pub fn standard_output_eravm(
 /// Runs the standard output mode for EVM.
 ///
 pub fn standard_output_evm(
-    input_files: &[PathBuf],
+    paths: &[PathBuf],
     libraries: Vec<String>,
-    solc: &mut SolcCompiler,
+    solc_compiler: &mut SolcCompiler,
     evm_version: Option<era_compiler_common::EVMVersion>,
     solc_optimizer_enabled: bool,
     optimizer_settings: era_compiler_llvm_context::OptimizerSettings,
@@ -350,13 +327,13 @@ pub fn standard_output_evm(
     remappings: Option<BTreeSet<String>>,
     debug_config: Option<era_compiler_llvm_context::DebugConfig>,
 ) -> anyhow::Result<EVMBuild> {
-    let solc_version = solc.version()?;
+    let solc_version = solc_compiler.version()?;
     let solc_pipeline = SolcPipeline::new(&solc_version, force_evmla);
 
     let solc_input = SolcStandardJsonInput::try_from_paths(
         SolcStandardJsonInputLanguage::Solidity,
         evm_version,
-        input_files,
+        paths,
         libraries,
         remappings,
         SolcStandardJsonInputSettingsSelection::new_required(Some(solc_pipeline)),
@@ -385,7 +362,7 @@ pub fn standard_output_evm(
     }
 
     let libraries = solc_input.settings.libraries.clone().unwrap_or_default();
-    let mut solc_output = solc.standard_json(
+    let mut solc_output = solc_compiler.standard_json(
         solc_input,
         Some(solc_pipeline),
         base_path,
@@ -409,11 +386,12 @@ pub fn standard_output_evm(
         }
     }
 
-    let project = solc_output.try_to_project_solidity(
+    let project = Project::try_from_solidity_sources(
+        &mut solc_output,
         source_code_files,
         libraries,
         solc_pipeline,
-        &solc_version,
+        solc_compiler,
         debug_config.as_ref(),
     )?;
 
@@ -426,7 +404,7 @@ pub fn standard_output_evm(
 /// Runs the standard JSON mode for EVM.
 ///
 pub fn standard_json_eravm(
-    solc: Option<&mut SolcCompiler>,
+    solc_compiler: Option<&mut SolcCompiler>,
     json_path: Option<PathBuf>,
     detect_missing_libraries: bool,
     force_evmla: bool,
@@ -441,12 +419,12 @@ pub fn standard_json_eravm(
     let solc_input = SolcStandardJsonInput::try_from_reader(json_path.as_deref())?;
     let language = solc_input.language;
 
-    let mut source_code_files = BTreeMap::new();
+    let mut sources = BTreeMap::new();
     for (path, source) in solc_input.sources.clone().into_iter() {
         let source: String = source
             .try_into()
             .map_err(|error| anyhow::anyhow!("Source `{path}` error: {error}"))?;
-        source_code_files.insert(path.to_owned(), source);
+        sources.insert(path.to_owned(), source);
     }
 
     let optimizer_settings =
@@ -459,12 +437,12 @@ pub fn standard_json_eravm(
     };
     let libraries = solc_input.settings.libraries.clone().unwrap_or_default();
 
-    let (mut solc_output, solc_version, project) = match (language, solc) {
-        (SolcStandardJsonInputLanguage::Solidity, Some(solc)) => {
-            let solc_version = solc.version()?;
+    let (mut solc_output, solc_version, project) = match (language, solc_compiler) {
+        (SolcStandardJsonInputLanguage::Solidity, Some(solc_compiler)) => {
+            let solc_version = solc_compiler.version()?;
             let solc_pipeline = SolcPipeline::new(&solc_version, force_evmla);
 
-            let mut solc_output = solc.standard_json(
+            let mut solc_output = solc_compiler.standard_json(
                 solc_input,
                 Some(solc_pipeline),
                 base_path,
@@ -485,11 +463,12 @@ pub fn standard_json_eravm(
                 std::process::exit(era_compiler_common::EXIT_CODE_SUCCESS);
             }
 
-            let project = solc_output.try_to_project_solidity(
-                source_code_files,
+            let project = Project::try_from_solidity_sources(
+                &mut solc_output,
+                sources,
                 libraries,
                 solc_pipeline,
-                &solc_version,
+                solc_compiler,
                 debug_config.as_ref(),
             )?;
 
@@ -498,17 +477,14 @@ pub fn standard_json_eravm(
         (SolcStandardJsonInputLanguage::Solidity, None) => {
             anyhow::bail!("Compiling Solidity without `solc` is not supported")
         }
-        (SolcStandardJsonInputLanguage::Yul, Some(solc)) => {
-            let solc_version = solc.version()?;
-            if solc.version()?.default != SolcCompiler::LAST_SUPPORTED_VERSION {
-                anyhow::bail!(
-                    "The Yul mode is only supported with the latest supported version of the Solidity compiler: {}",
-                    SolcCompiler::LAST_SUPPORTED_VERSION,
-                );
-            }
-
-            let mut solc_output =
-                solc.standard_json(solc_input, None, base_path, include_paths, allow_paths)?;
+        (SolcStandardJsonInputLanguage::Yul, Some(solc_compiler)) => {
+            let solc_output = solc_compiler.standard_json(
+                solc_input,
+                None,
+                base_path,
+                include_paths,
+                allow_paths,
+            )?;
             if solc_output
                 .errors
                 .as_deref()
@@ -523,21 +499,36 @@ pub fn standard_json_eravm(
                 std::process::exit(era_compiler_common::EXIT_CODE_SUCCESS);
             }
 
-            let project = solc_output.try_to_project_yul(
-                source_code_files,
+            let project = Project::try_from_yul_sources(
+                sources,
                 libraries,
-                &solc_version,
+                Some(solc_compiler),
                 debug_config.as_ref(),
             )?;
 
-            (solc_output, Some(solc_version), project)
+            (solc_output, Some(solc_compiler.version()?), project)
         }
-        (SolcStandardJsonInputLanguage::Yul, None) => todo!(),
-        (SolcStandardJsonInputLanguage::LLVMIR, _) => {
-            anyhow::bail!("Compiling LLVM IR in standard JSON mode is not supported yet")
+        (SolcStandardJsonInputLanguage::Yul, None) => {
+            let solc_output = SolcStandardJsonOutput::new(&sources);
+            let project =
+                Project::try_from_yul_sources(sources, libraries, None, debug_config.as_ref())?;
+            (solc_output, None, project)
         }
-        (SolcStandardJsonInputLanguage::EraVMAssembly, _) => {
-            anyhow::bail!("Compiling EraVM assembly in standard JSON mode is not supported yet")
+        (SolcStandardJsonInputLanguage::LLVMIR, Some(_)) => {
+            anyhow::bail!("LLVM IR projects cannot be compiled with `solc`")
+        }
+        (SolcStandardJsonInputLanguage::LLVMIR, None) => {
+            let solc_output = SolcStandardJsonOutput::new(&sources);
+            let project = Project::try_from_llvm_ir_sources(sources)?;
+            (solc_output, None, project)
+        }
+        (SolcStandardJsonInputLanguage::EraVMAssembly, Some(_)) => {
+            anyhow::bail!("EraVM assembly projects cannot be compiled with `solc`")
+        }
+        (SolcStandardJsonInputLanguage::EraVMAssembly, None) => {
+            let solc_output = SolcStandardJsonOutput::new(&sources);
+            let project = Project::try_from_eravm_assembly_sources(sources)?;
+            (solc_output, None, project)
         }
     };
 
@@ -566,7 +557,7 @@ pub fn standard_json_eravm(
 /// Runs the standard JSON mode for EVM.
 ///
 pub fn standard_json_evm(
-    solc: Option<&mut SolcCompiler>,
+    solc_compiler: Option<&mut SolcCompiler>,
     json_path: Option<PathBuf>,
     force_evmla: bool,
     base_path: Option<String>,
@@ -579,12 +570,12 @@ pub fn standard_json_evm(
     let solc_input = SolcStandardJsonInput::try_from_reader(json_path.as_deref())?;
     let language = solc_input.language;
 
-    let mut source_code_files = BTreeMap::new();
+    let mut sources = BTreeMap::new();
     for (path, source) in solc_input.sources.clone().into_iter() {
         let source: String = source
             .try_into()
             .map_err(|error| anyhow::anyhow!("Source `{path}` error: {error}"))?;
-        source_code_files.insert(path.to_owned(), source);
+        sources.insert(path.to_owned(), source);
     }
 
     let optimizer_settings =
@@ -597,11 +588,11 @@ pub fn standard_json_evm(
     };
     let libraries = solc_input.settings.libraries.clone().unwrap_or_default();
 
-    let (mut solc_output, solc_version, project) = match (language, solc) {
-        (SolcStandardJsonInputLanguage::Solidity, Some(solc)) => {
-            let solc_version = solc.version()?;
+    let (mut solc_output, solc_version, project) = match (language, solc_compiler) {
+        (SolcStandardJsonInputLanguage::Solidity, Some(solc_compiler)) => {
+            let solc_version = solc_compiler.version()?;
             let solc_pipeline = SolcPipeline::new(&solc_version, force_evmla);
-            let mut solc_output = solc.standard_json(
+            let mut solc_output = solc_compiler.standard_json(
                 solc_input,
                 Some(solc_pipeline),
                 base_path,
@@ -622,11 +613,12 @@ pub fn standard_json_evm(
                 std::process::exit(era_compiler_common::EXIT_CODE_SUCCESS);
             }
 
-            let project = solc_output.try_to_project_solidity(
-                source_code_files,
+            let project = Project::try_from_solidity_sources(
+                &mut solc_output,
+                sources,
                 libraries,
                 solc_pipeline,
-                &solc_version,
+                solc_compiler,
                 debug_config.as_ref(),
             )?;
 
@@ -635,17 +627,14 @@ pub fn standard_json_evm(
         (SolcStandardJsonInputLanguage::Solidity, None) => {
             anyhow::bail!("Compiling Solidity without `solc` is not supported")
         }
-        (SolcStandardJsonInputLanguage::Yul, Some(solc)) => {
-            let solc_version = solc.version()?;
-            if solc.version()?.default != SolcCompiler::LAST_SUPPORTED_VERSION {
-                anyhow::bail!(
-                    "The Yul mode is only supported with the latest supported version of the Solidity compiler: {}",
-                    SolcCompiler::LAST_SUPPORTED_VERSION,
-                );
-            }
-
-            let mut solc_output =
-                solc.standard_json(solc_input, None, base_path, include_paths, allow_paths)?;
+        (SolcStandardJsonInputLanguage::Yul, Some(solc_compiler)) => {
+            let solc_output = solc_compiler.standard_json(
+                solc_input,
+                None,
+                base_path,
+                include_paths,
+                allow_paths,
+            )?;
             if solc_output
                 .errors
                 .as_deref()
@@ -660,18 +649,28 @@ pub fn standard_json_evm(
                 std::process::exit(era_compiler_common::EXIT_CODE_SUCCESS);
             }
 
-            let project = solc_output.try_to_project_yul(
-                source_code_files,
+            let project = Project::try_from_yul_sources(
+                sources,
                 libraries,
-                &solc_version,
+                Some(solc_compiler),
                 debug_config.as_ref(),
             )?;
 
-            (solc_output, Some(solc_version), project)
+            (solc_output, Some(solc_compiler.version()?), project)
         }
-        (SolcStandardJsonInputLanguage::Yul, None) => todo!(),
-        (SolcStandardJsonInputLanguage::LLVMIR, _) => {
-            anyhow::bail!("Compiling LLVM IR in standard JSON mode is not supported yet")
+        (SolcStandardJsonInputLanguage::Yul, None) => {
+            let solc_output = SolcStandardJsonOutput::new(&sources);
+            let project =
+                Project::try_from_yul_sources(sources, libraries, None, debug_config.as_ref())?;
+            (solc_output, None, project)
+        }
+        (SolcStandardJsonInputLanguage::LLVMIR, Some(_)) => {
+            anyhow::bail!("LLVM IR projects cannot be compiled with `solc`")
+        }
+        (SolcStandardJsonInputLanguage::LLVMIR, None) => {
+            let solc_output = SolcStandardJsonOutput::new(&sources);
+            let project = Project::try_from_llvm_ir_sources(sources)?;
+            (solc_output, None, project)
         }
         (SolcStandardJsonInputLanguage::EraVMAssembly, _) => {
             anyhow::bail!("Compiling EraVM assembly to EVM is not supported")
@@ -689,9 +688,9 @@ pub fn standard_json_evm(
 ///
 pub fn combined_json_eravm(
     format: String,
-    input_files: &[PathBuf],
+    paths: &[PathBuf],
     libraries: Vec<String>,
-    solc: &mut SolcCompiler,
+    solc_compiler: &mut SolcCompiler,
     evm_version: Option<era_compiler_common::EVMVersion>,
     solc_optimizer_enabled: bool,
     optimizer_settings: era_compiler_llvm_context::OptimizerSettings,
@@ -711,9 +710,9 @@ pub fn combined_json_eravm(
     let zksolc_version = semver::Version::parse(env!("CARGO_PKG_VERSION")).expect("Always valid");
 
     let build = standard_output_eravm(
-        input_files,
+        paths,
         libraries,
-        solc,
+        solc_compiler,
         evm_version,
         solc_optimizer_enabled,
         optimizer_settings,
@@ -729,7 +728,7 @@ pub fn combined_json_eravm(
         debug_config,
     )?;
 
-    let mut combined_json = solc.combined_json(input_files, format.as_str())?;
+    let mut combined_json = solc_compiler.combined_json(paths, format.as_str())?;
     build.write_to_combined_json(&mut combined_json, &zksolc_version)?;
 
     match output_directory {
@@ -754,9 +753,9 @@ pub fn combined_json_eravm(
 ///
 pub fn combined_json_evm(
     format: String,
-    input_files: &[PathBuf],
+    paths: &[PathBuf],
     libraries: Vec<String>,
-    solc: &mut SolcCompiler,
+    solc_compiler: &mut SolcCompiler,
     evm_version: Option<era_compiler_common::EVMVersion>,
     solc_optimizer_enabled: bool,
     optimizer_settings: era_compiler_llvm_context::OptimizerSettings,
@@ -774,9 +773,9 @@ pub fn combined_json_evm(
     let zksolc_version = semver::Version::parse(env!("CARGO_PKG_VERSION")).expect("Always valid");
 
     let build = standard_output_evm(
-        input_files,
+        paths,
         libraries,
-        solc,
+        solc_compiler,
         evm_version,
         solc_optimizer_enabled,
         optimizer_settings,
@@ -790,7 +789,7 @@ pub fn combined_json_evm(
         debug_config,
     )?;
 
-    let mut combined_json = solc.combined_json(input_files, format.as_str())?;
+    let mut combined_json = solc_compiler.combined_json(paths, format.as_str())?;
     build.write_to_combined_json(&mut combined_json, &zksolc_version)?;
 
     match output_directory {