+++ src/project/mod.rs
@@ -7,7 +7,7 @@ pub mod contract;
 use std::collections::BTreeMap;
 use std::collections::HashMap;
 use std::collections::HashSet;
-use std::path::Path;
+use std::path::PathBuf;
 
 use rayon::iter::IntoParallelIterator;
 use rayon::iter::ParallelIterator;
@@ -24,8 +24,12 @@ use crate::process::input_eravm::Input as EraVMProcessInput;
 use crate::process::input_evm::Input as EVMProcessInput;
 use crate::process::output_eravm::Output as EraVMProcessOutput;
 use crate::process::output_evm::Output as EVMProcessOutput;
+use crate::project::contract::ir::IR as ProjectContractIR;
 use crate::project::contract::ir::IR;
+use crate::project::contract::Contract as ProjectContract;
+use crate::solc::pipeline::Pipeline as SolcPipeline;
 use crate::solc::standard_json::input::language::Language as SolcStandardJsonInputLanguage;
+use crate::solc::standard_json::output::Output as SolcStandardJsonOutput;
 use crate::solc::version::Version as SolcVersion;
 use crate::solc::Compiler as SolcCompiler;
 use crate::yul::lexer::Lexer;
@@ -40,8 +44,8 @@ use self::contract::Contract;
 pub struct Project {
     /// The project language.
     pub language: SolcStandardJsonInputLanguage,
-    /// The source code version.
-    pub version: SolcVersion,
+    /// The `solc` compiler version.
+    pub solc_version: Option<SolcVersion>,
     /// The project contracts,
     pub contracts: BTreeMap<String, Contract>,
     /// The mapping of auxiliary identifiers, e.g. Yul object names, to full contract paths.
@@ -56,7 +60,7 @@ impl Project {
     ///
     pub fn new(
         language: SolcStandardJsonInputLanguage,
-        version: SolcVersion,
+        solc_version: Option<SolcVersion>,
         contracts: BTreeMap<String, Contract>,
         libraries: BTreeMap<String, BTreeMap<String, String>>,
     ) -> Self {
@@ -67,11 +71,282 @@ impl Project {
 
         Self {
             language,
-            version,
+            solc_version,
             contracts,
             identifier_paths,
             libraries,
         }
+    }
+
+    ///
+    /// Parses the Solidity `sources` and returns a Solidity project.
+    ///
+    pub fn try_from_solidity_sources(
+        solc_output: &mut SolcStandardJsonOutput,
+        source_code_files: BTreeMap<String, String>,
+        libraries: BTreeMap<String, BTreeMap<String, String>>,
+        pipeline: SolcPipeline,
+        solc_compiler: &mut SolcCompiler,
+        debug_config: Option<&era_compiler_llvm_context::DebugConfig>,
+    ) -> anyhow::Result<Self> {
+        if let SolcPipeline::EVMLA = pipeline {
+            solc_output.preprocess_dependencies()?;
+        }
+
+        let files = match solc_output.contracts.as_ref() {
+            Some(files) => files,
+            None => {
+                anyhow::bail!(
+                    "{}",
+                    solc_output
+                        .errors
+                        .as_ref()
+                        .map(|errors| serde_json::to_string_pretty(errors).expect("Always valid"))
+                        .unwrap_or_else(|| "Unknown project assembling error".to_owned())
+                );
+            }
+        };
+        let mut project_contracts = BTreeMap::new();
+
+        let solc_version = solc_compiler.version()?;
+
+        for (path, contracts) in files.iter() {
+            for (name, contract) in contracts.iter() {
+                let full_path = format!("{path}:{name}");
+
+                let source = match pipeline {
+                    SolcPipeline::Yul => {
+                        let ir_optimized = match contract.ir_optimized.to_owned() {
+                            Some(ir_optimized) => ir_optimized,
+                            None => continue,
+                        };
+                        if ir_optimized.is_empty() {
+                            continue;
+                        }
+
+                        if let Some(debug_config) = debug_config {
+                            debug_config.dump_yul(
+                                full_path.as_str(),
+                                None,
+                                ir_optimized.as_str(),
+                            )?;
+                        }
+
+                        let mut lexer = Lexer::new(ir_optimized.to_owned());
+                        let object = Object::parse(&mut lexer, None).map_err(|error| {
+                            anyhow::anyhow!("Contract `{}` parsing error: {:?}", full_path, error)
+                        })?;
+
+                        ProjectContractIR::new_yul(ir_optimized.to_owned(), object)
+                    }
+                    SolcPipeline::EVMLA => {
+                        let evm = contract.evm.as_ref();
+                        let assembly = match evm.and_then(|evm| evm.assembly.to_owned()) {
+                            Some(assembly) => assembly.to_owned(),
+                            None => continue,
+                        };
+                        let extra_metadata = evm
+                            .and_then(|evm| evm.extra_metadata.to_owned())
+                            .unwrap_or_default();
+
+                        ProjectContractIR::new_evmla(assembly, extra_metadata)
+                    }
+                };
+
+                let source_code = source_code_files
+                    .get(path.as_str())
+                    .ok_or_else(|| anyhow::anyhow!("Source code for path `{}` not found", path))?;
+                let hash = sha3::Keccak256::digest(source_code.as_bytes()).into();
+
+                let project_contract = ProjectContract::new(
+                    full_path.clone(),
+                    source,
+                    contract.metadata.to_owned(),
+                    hash,
+                    Some(&solc_version),
+                );
+                project_contracts.insert(full_path, project_contract);
+            }
+        }
+
+        Ok(Project::new(
+            SolcStandardJsonInputLanguage::Solidity,
+            Some(solc_version.to_owned()),
+            project_contracts,
+            libraries,
+        ))
+    }
+
+    ///
+    /// Reads the Yul source code `paths` and returns a Yul project.
+    ///
+    pub fn try_from_yul_paths(
+        paths: &[PathBuf],
+        libraries: BTreeMap<String, BTreeMap<String, String>>,
+        solc_compiler: Option<&mut SolcCompiler>,
+        debug_config: Option<&era_compiler_llvm_context::DebugConfig>,
+    ) -> anyhow::Result<Self> {
+        let sources = paths
+            .iter()
+            .map(|path| {
+                let source_code = std::fs::read_to_string(path.as_path())
+                    .map_err(|error| anyhow::anyhow!("Yul file {path:?} reading error: {error}"))?;
+                Ok((path.to_string_lossy().to_string(), source_code))
+            })
+            .collect::<anyhow::Result<BTreeMap<String, String>>>()?;
+        Self::try_from_yul_sources(sources, libraries, solc_compiler, debug_config)
+    }
+
+    ///
+    /// Parses the Yul `sources` and returns a Yul project.
+    ///
+    pub fn try_from_yul_sources(
+        sources: BTreeMap<String, String>,
+        libraries: BTreeMap<String, BTreeMap<String, String>>,
+        solc_compiler: Option<&mut SolcCompiler>,
+        debug_config: Option<&era_compiler_llvm_context::DebugConfig>,
+    ) -> anyhow::Result<Self> {
+        let solc_version = match solc_compiler {
+            Some(solc) => {
+                if solc.version()?.default != SolcCompiler::LAST_SUPPORTED_VERSION {
+                    anyhow::bail!(
+                        "The Yul mode is only supported with the latest supported version of the Solidity compiler: {}",
+                        SolcCompiler::LAST_SUPPORTED_VERSION,
+                    );
+                }
+
+                for path in sources.keys() {
+                    solc.validate_yul(PathBuf::from(path).as_path())?;
+                }
+                Some(solc.version()?)
+            }
+            None => None,
+        };
+
+        let project_contracts = sources
+            .into_par_iter()
+            .map(|(path, source_code)| {
+                let hash = sha3::Keccak256::digest(source_code.as_bytes()).into();
+
+                let mut lexer = Lexer::new(source_code.to_owned());
+                let object = Object::parse(&mut lexer, None).map_err(|error| {
+                    anyhow::anyhow!("Yul object `{}` parsing error: {}", path, error)
+                })?;
+
+                if let Some(debug_config) = debug_config {
+                    debug_config.dump_yul(path.as_str(), None, source_code.as_str())?;
+                }
+
+                let contract = Contract::new(
+                    path.clone(),
+                    IR::new_yul(source_code.to_owned(), object),
+                    None,
+                    hash,
+                    solc_version.as_ref(),
+                );
+
+                Ok((path, contract))
+            })
+            .collect::<anyhow::Result<BTreeMap<String, Contract>>>()?;
+
+        Ok(Self::new(
+            SolcStandardJsonInputLanguage::Yul,
+            solc_version,
+            project_contracts,
+            libraries,
+        ))
+    }
+
+    ///
+    /// Reads the LLVM IR source code `paths` and returns an LLVM IR project.
+    ///
+    pub fn try_from_llvm_ir_paths(paths: &[PathBuf]) -> anyhow::Result<Self> {
+        let sources = paths
+            .iter()
+            .map(|path| {
+                let source_code = std::fs::read_to_string(path.as_path()).map_err(|error| {
+                    anyhow::anyhow!("LLVM IR file {path:?} reading error: {error}")
+                })?;
+                Ok((path.to_string_lossy().to_string(), source_code))
+            })
+            .collect::<anyhow::Result<BTreeMap<String, String>>>()?;
+        Self::try_from_llvm_ir_sources(sources)
+    }
+
+    ///
+    /// Parses the LLVM IR `sources` and returns an LLVM IR project.
+    ///
+    pub fn try_from_llvm_ir_sources(sources: BTreeMap<String, String>) -> anyhow::Result<Self> {
+        let project_contracts = sources
+            .into_par_iter()
+            .map(|(path, source_code)| {
+                let hash = sha3::Keccak256::digest(source_code.as_bytes()).into();
+
+                let contract = Contract::new(
+                    path.clone(),
+                    IR::new_llvm_ir(path.clone(), source_code),
+                    None,
+                    hash,
+                    None,
+                );
+
+                Ok((path, contract))
+            })
+            .collect::<anyhow::Result<BTreeMap<String, Contract>>>()?;
+
+        Ok(Self::new(
+            SolcStandardJsonInputLanguage::LLVMIR,
+            None,
+            project_contracts,
+            BTreeMap::new(),
+        ))
+    }
+
+    ///
+    /// Reads the EraVM assembly source code `paths` and returns an EraVM assembly project.
+    ///
+    pub fn try_from_eravm_assembly_paths(paths: &[PathBuf]) -> anyhow::Result<Self> {
+        let sources = paths
+            .iter()
+            .map(|path| {
+                let source_code = std::fs::read_to_string(path.as_path()).map_err(|error| {
+                    anyhow::anyhow!("EraVM assembly file {path:?} reading error: {error}")
+                })?;
+                Ok((path.to_string_lossy().to_string(), source_code))
+            })
+            .collect::<anyhow::Result<BTreeMap<String, String>>>()?;
+        Self::try_from_eravm_assembly_sources(sources)
+    }
+
+    ///
+    /// Parses the EraVM assembly `sources` and returns an EraVM assembly project.
+    ///
+    pub fn try_from_eravm_assembly_sources(
+        sources: BTreeMap<String, String>,
+    ) -> anyhow::Result<Self> {
+        let project_contracts = sources
+            .into_par_iter()
+            .map(|(path, source_code)| {
+                let hash = sha3::Keccak256::digest(source_code.as_bytes()).into();
+
+                let contract = Contract::new(
+                    path.clone(),
+                    IR::new_eravm_assembly(path.clone(), source_code),
+                    None,
+                    hash,
+                    None,
+                );
+
+                Ok((path, contract))
+            })
+            .collect::<anyhow::Result<BTreeMap<String, Contract>>>()?;
+
+        Ok(Self::new(
+            SolcStandardJsonInputLanguage::EraVMAssembly,
+            None,
+            project_contracts,
+            BTreeMap::new(),
+        ))
     }
 
     ///
@@ -223,124 +498,6 @@ impl Project {
             missing_deployable_libraries.insert(contract_path.to_owned(), missing_libraries);
         }
         MissingLibraries::new(missing_deployable_libraries)
-    }
-
-    ///
-    /// Parses the Yul source code file and returns the source data.
-    ///
-    pub fn try_from_yul_path(
-        path: &Path,
-        solc_validator: Option<&SolcCompiler>,
-    ) -> anyhow::Result<Self> {
-        let source_code = std::fs::read_to_string(path)
-            .map_err(|error| anyhow::anyhow!("Yul file {:?} reading error: {}", path, error))?;
-        Self::try_from_yul_string(path, source_code.as_str(), solc_validator)
-    }
-
-    ///
-    /// Parses the test Yul source code string and returns the source data.
-    ///
-    /// Only for integration testing purposes.
-    ///
-    pub fn try_from_yul_string(
-        path: &Path,
-        source_code: &str,
-        solc_validator: Option<&SolcCompiler>,
-    ) -> anyhow::Result<Self> {
-        if let Some(solc) = solc_validator {
-            solc.validate_yul(path)?;
-        }
-
-        let source_version = SolcVersion::new_simple(SolcCompiler::LAST_SUPPORTED_VERSION);
-        let path = path.to_string_lossy().to_string();
-        let source_hash = sha3::Keccak256::digest(source_code.as_bytes()).into();
-
-        let mut lexer = Lexer::new(source_code.to_owned());
-        let object = Object::parse(&mut lexer, None)
-            .map_err(|error| anyhow::anyhow!("Yul object `{}` parsing error: {}", path, error))?;
-
-        let mut project_contracts = BTreeMap::new();
-        project_contracts.insert(
-            path.to_owned(),
-            Contract::new(
-                path,
-                source_hash,
-                source_version.clone(),
-                IR::new_yul(source_code.to_owned(), object),
-                None,
-            ),
-        );
-
-        Ok(Self::new(
-            SolcStandardJsonInputLanguage::Yul,
-            source_version,
-            project_contracts,
-            BTreeMap::new(),
-        ))
-    }
-
-    ///
-    /// Parses the LLVM IR source code file and returns the source data.
-    ///
-    pub fn try_from_llvm_ir_path(path: &Path) -> anyhow::Result<Self> {
-        let source_code = std::fs::read_to_string(path)
-            .map_err(|error| anyhow::anyhow!("LLVM IR file {:?} reading error: {}", path, error))?;
-        let source_hash = sha3::Keccak256::digest(source_code.as_bytes()).into();
-
-        let source_version = SolcVersion::new_simple(era_compiler_llvm_context::LLVM_VERSION);
-        let path = path.to_string_lossy().to_string();
-
-        let mut project_contracts = BTreeMap::new();
-        project_contracts.insert(
-            path.clone(),
-            Contract::new(
-                path.clone(),
-                source_hash,
-                source_version.clone(),
-                IR::new_llvm_ir(path, source_code),
-                None,
-            ),
-        );
-
-        Ok(Self::new(
-            SolcStandardJsonInputLanguage::LLVMIR,
-            source_version,
-            project_contracts,
-            BTreeMap::new(),
-        ))
-    }
-
-    ///
-    /// Parses the EraVM assembly source code file and returns the source data.
-    ///
-    pub fn try_from_eravm_assembly_path(path: &Path) -> anyhow::Result<Self> {
-        let source_code = std::fs::read_to_string(path).map_err(|error| {
-            anyhow::anyhow!("EraVM assembly file {:?} reading error: {}", path, error)
-        })?;
-        let source_hash = sha3::Keccak256::digest(source_code.as_bytes()).into();
-
-        let source_version =
-            SolcVersion::new_simple(era_compiler_llvm_context::eravm_const::ZKEVM_VERSION);
-        let path = path.to_string_lossy().to_string();
-
-        let mut project_contracts = BTreeMap::new();
-        project_contracts.insert(
-            path.clone(),
-            Contract::new(
-                path.clone(),
-                source_hash,
-                source_version.clone(),
-                IR::new_zkasm(path, source_code),
-                None,
-            ),
-        );
-
-        Ok(Self::new(
-            SolcStandardJsonInputLanguage::EraVMAssembly,
-            source_version,
-            project_contracts,
-            BTreeMap::new(),
-        ))
     }
 }